+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./App.css";
-import { CardList } from "./CardList";
-import { Card } from "./Card"
+import { CardList } from "./Components/CardList";
+import { Card } from "./Components/Card";
 
 //import { CardList } from "./CardList"
